backend/requirements.txt
Django==4.2.22
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Pillow==10.2.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-storages==1.14.2
gunicorn==21.2.0
whitenoise==6.6.0 